Fix tests after moving modules in src/vocalpy/
